Inventory UI should be all hooked up READ DESCRIPTION
The main issue / task is making canvas persist through scenes. I have the code for DontDestroyOnLoad but the canvas in Dead's greybox and other scenes we are transitioning to will need to have a scene setter script to add any needed components they need on the canvas. They will likely need to be disabled or removed when leaving the scene.